Move sociogram related components into self contained module
