+++ src/containers/Elements/Sociogram/LayoutNode.js
@@ -4,9 +4,9 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { first } from 'lodash';
 import { Node } from 'network-canvas-ui';
-import { selectable } from '../behaviours';
-import { actionCreators as networkActions } from '../ducks/modules/network';
-import { DragSource } from '../behaviours/DragAndDrop';
+import { selectable } from '../../../behaviours';
+import { actionCreators as networkActions } from '../../../ducks/modules/network';
+import { DragSource } from '../../../behaviours/DragAndDrop';
 
 const label = node => node.nickname;
