+++ src/containers/Elements/Sociogram/index.js
@@ -1,0 +1,3 @@
+import Sociogram from './Sociogram';
+
+export default Sociogram;
