+++ src/containers/Elements/index.js
@@ -5,8 +5,5 @@ export { default as Form } from './Form';
 export { default as FormWizard } from './FormWizard';
 export { default as NodeForm } from './NodeForm';
 export { default as Field } from './Field';
-export { default as EdgeLayout } from './EdgeLayout';
-export { default as NodeLayout } from './NodeLayout';
 export { default as NodeBucket } from './NodeBucket';
-export { default as SociogramRadar } from './SociogramRadar';
-export { default as SociogramBackground } from './SociogramBackground';
+export { default as Sociogram } from './Sociogram';
